+++ src/constants/BarrelFiles.ts
@@ -1,1 +1,0 @@
-export const BarrelFiles = ["index.ts", "index.tsx"];