+++ src/helpers/getFileContents.ts
@@ -1,13 +1,13 @@
 import { join } from "path";
-import { fileExists, readFile } from ".";
+import { fileExists, getBarrelFiles, readFile } from ".";
 import { FileType, Uri, workspace } from "vscode";
-import { BarrelFiles } from "../constants";
 
 export const getFileContents = async (
   filePath: string
 ): Promise<string | undefined> => {
   const stat = await workspace.fs.stat(Uri.file(filePath));
   if (stat.type === FileType.Directory) {
+    const BarrelFiles = getBarrelFiles();
     for (const indexFile of BarrelFiles) {
       const indexPath = join(filePath, indexFile);
       if (await fileExists(indexPath)) {