+++ src/models/FileOrFolderToExport.ts
@@ -1,4 +1,6 @@
 export interface FileOrFolderToExport {
   name: string;
+  absPath: string;
+  relPath: string;
   type: "file" | "folder";
 }