+++ src/providers/ExportProvider.ts
@@ -1,14 +1,32 @@
-import * as vscode from 'vscode';
-import * as path from 'path';
-import * as fs from 'fs';
-import { EXTENSION_KEY, CONFIG_FOLDERS, CONFIG_RELATIVE_EXCLUDE } from '../constants';
-import { ExportAll } from '../commands';
-import { pathResolve, getAbsoluteFolderPath } from '../helpers';
+import * as vscode from "vscode";
+import * as path from "path";
+import * as fs from "fs";
+import {
+  EXTENSION_KEY,
+  CONFIG_FOLDERS,
+  CONFIG_RELATIVE_EXCLUDE,
+  CONFIG_BARREL_NAME,
+} from "../constants";
+import { ExportAll } from "../commands";
+import { pathResolve, getAbsoluteFolderPath } from "../helpers";
+import { workspace } from "vscode";
 
 export class ExportFolder extends vscode.TreeItem {
-  
-  constructor(public label: string, public value: string | null, public uri?: vscode.Uri | null, command?: vscode.Command | null, public iconPath?: string | vscode.ThemeIcon, public contextValue?: string, public children?: vscode.TreeItem[]) {
-    super(label, children ? vscode.TreeItemCollapsibleState.Expanded : vscode.TreeItemCollapsibleState.None);
+  constructor(
+    public label: string,
+    public value: string | null,
+    public uri?: vscode.Uri | null,
+    command?: vscode.Command | null,
+    public iconPath?: string | vscode.ThemeIcon,
+    public contextValue?: string,
+    public children?: vscode.TreeItem[]
+  ) {
+    super(
+      label,
+      children
+        ? vscode.TreeItemCollapsibleState.Expanded
+        : vscode.TreeItemCollapsibleState.None
+    );
 
     if (command) {
       this.command = command;
@@ -22,11 +40,13 @@ export class ExportFolder extends vscode.TreeItem {
 }
 
 export class ExportProvider implements vscode.TreeDataProvider<ExportFolder> {
-  private _onDidChangeTreeData: vscode.EventEmitter<ExportFolder | undefined> = new vscode.EventEmitter<ExportFolder | undefined>();
-  readonly onDidChangeTreeData: vscode.Event<ExportFolder | undefined> = this._onDidChangeTreeData.event;
+  private _onDidChangeTreeData: vscode.EventEmitter<ExportFolder | undefined> =
+    new vscode.EventEmitter<ExportFolder | undefined>();
+  readonly onDidChangeTreeData: vscode.Event<ExportFolder | undefined> =
+    this._onDidChangeTreeData.event;
 
   constructor() {
-    console.log('Refresh');
+    console.log("Refresh");
   }
 
   public getTreeItem(folder: ExportFolder): vscode.TreeItem {
@@ -35,7 +55,7 @@ export class ExportProvider implements vscode.TreeDataProvider<ExportFolder> {
 
   /**
    * Retrieve the children to show in the tree view
-   * @param element 
+   * @param element
    */
   public async getChildren(element?: ExportFolder): Promise<ExportFolder[]> {
     if (element && element.children) {
@@ -50,12 +70,24 @@ export class ExportProvider implements vscode.TreeDataProvider<ExportFolder> {
     const excludeList = await this.getExportItems(excludeOptions, true);
 
     return [
-      new ExportFolder(`Folder listeners`, null, null, null, '', 'exportall.header', [
-        ...exportListeners
-      ]),
-      new ExportFolder(`Excluded folders & files`, null, null, null, '', 'exportall.header', [
-        ...excludeList
-      ])
+      new ExportFolder(
+        `Folder listeners`,
+        null,
+        null,
+        null,
+        "",
+        "exportall.header",
+        [...exportListeners]
+      ),
+      new ExportFolder(
+        `Excluded folders & files`,
+        null,
+        null,
+        null,
+        "",
+        "exportall.header",
+        [...excludeList]
+      ),
     ];
   }
 
@@ -68,21 +100,27 @@ export class ExportProvider implements vscode.TreeDataProvider<ExportFolder> {
 
   /**
    * Get all the items for the tree view
-   * @param options 
+   * @param options
    */
-  public async getExportItems(options: string[], skipIndexCheck: boolean = false) {
+  public async getExportItems(
+    options: string[],
+    skipIndexCheck: boolean = false
+  ) {
     let items: ExportFolder[] = [];
     for (const opt of options) {
       try {
+        const wsConfig = workspace.getConfiguration(EXTENSION_KEY);
+        const barrelName = wsConfig.get<string>(CONFIG_BARREL_NAME);
+
         const absPath = getAbsoluteFolderPath(opt);
         const isFolder = fs.lstatSync(absPath).isDirectory();
         const uri = vscode.Uri.file(absPath);
-        const indexPath = path.join(absPath, 'index.ts');
+        const indexPath = path.join(absPath, barrelName || "index.ts");
 
         // When no `index.ts` file exists, create one
         if (!skipIndexCheck) {
           if (!fs.existsSync(indexPath)) {
-            await ExportAll.start(uri); 
+            await ExportAll.start(uri);
           }
         }
 
@@ -95,24 +133,31 @@ export class ExportProvider implements vscode.TreeDataProvider<ExportFolder> {
 
         // Create the open command
         let command: vscode.Command = {
-          command: 'exportall.open',
-          title: '',
-          arguments: [fileUri]
+          command: "exportall.open",
+          title: "",
+          arguments: [fileUri],
         };
-        
+
         const config = {
           label: path.basename(uri.fsPath),
           value: absPath,
           uri: uri,
           command,
           iconPath: isFolder ? vscode.ThemeIcon.Folder : vscode.ThemeIcon.File,
-          contextValue: skipIndexCheck ? "exportall.excluded" : "exportall.listener"
+          contextValue: skipIndexCheck
+            ? "exportall.excluded"
+            : "exportall.listener",
         };
-        const folder = new ExportFolder(config.label, config.value, config.uri, config.command, config.iconPath, config.contextValue);
+        const folder = new ExportFolder(
+          config.label,
+          config.value,
+          config.uri,
+          config.command,
+          config.iconPath,
+          config.contextValue
+        );
         items.push(folder);
-      } catch {
-
-      }
+      } catch {}
     }
     return items;
   }